+++ BackEnd/src/main/java/com/manhattan/busyness_predictor/service/FavouriteService.java
@@ -1,14 +1,15 @@
 package com.manhattan.busyness_predictor.service;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import com.manhattan.busyness_predictor.model.Favourite;
 import com.manhattan.busyness_predictor.model.Location;
 import com.manhattan.busyness_predictor.repository.FavouriteRepository;
 import com.manhattan.busyness_predictor.repository.LocationRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class FavouriteService {
@@ -19,7 +20,7 @@ public class FavouriteService {
     @Autowired
     private LocationRepository locationRepository;
 
-    public void addToFavourites(Long userId, Long locationId) {
+    public void addToFavourites(Integer userId, Integer locationId) {
         // Check if location exists
         locationRepository.findById(locationId)
                 .orElseThrow(() -> new RuntimeException("Location not found"));
@@ -33,14 +34,14 @@ public class FavouriteService {
         favouriteRepository.save(favourite);
     }
 
-    public void removeFromFavourites(Long userId, Long locationId) {
+    public void removeFromFavourites(Integer userId, Integer locationId) {
         Favourite favourite = favouriteRepository.findByUserIdAndLocationId(userId, locationId)
                 .orElseThrow(() -> new RuntimeException("Location is not in favourites"));
 
         favouriteRepository.delete(favourite);
     }
 
-    public List<Location> getFavouriteLocations(Long userId) {
+    public List<Location> getFavouriteLocations(Integer userId) {
         List<Favourite> favourites = favouriteRepository.findByUserId(userId);
 
         return favourites.stream()
@@ -50,7 +51,7 @@ public class FavouriteService {
                 .collect(Collectors.toList());
     }
 
-    public boolean isFavourited(Long userId, Long locationId) {
+    public boolean isFavourited(Integer userId, Integer locationId) {
         return favouriteRepository.existsByUserIdAndLocationId(userId, locationId);
     }
 }
